Create setup_machine.sh
setup_machine.sh contains commands that sets up the machine for benchmarking. More specifically, it disables turbo boost and CPU frequency scaling (and hyperthreading, but that is commented for now)

Needs sudo or root permissions in order to work.

The changes applied by this script are only temporary, and can be undone by restarting the PC.